Requirment Changes

+++ app/Http/Controllers/GalleryAllController.php
@@ -16,11 +16,14 @@ class GalleryAllController extends Controller
      */
     public function index()
     {
-        $galleries = GalleryAll::latest()->paginate(10);
+        if (auth()->user()->can('Gallery View')) {
+            $galleries = GalleryAll::latest()->paginate(10);
 
-        return view('backend.gallery.index', [
-            'galleries' => $galleries,
-        ]);
+            return view('backend.gallery.index', [
+                'galleries' => $galleries,
+            ]);
+        }
+        abort('404');
     }
 
     /**
@@ -30,8 +33,10 @@ class GalleryAllController extends Controller
      */
     public function create()
     {
-
-        return view('backend.gallery.create');
+        if (auth()->user()->can('Gallery Create')) {
+            return view('backend.gallery.create');
+        }
+        abort('404');
     }
 
     /**
@@ -42,42 +47,41 @@ class GalleryAllController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        if (auth()->user()->can('Gallery Create')) {
+            $request->validate([
 
 
-            'image' => ['mimes:png,jpeg,jpg',],
+                'image' => ['mimes:png,jpeg,jpg',],
+
+            ], [
+                'image.*.mimes' => '  Image must be png,jpg,jpeg formate',
+
+            ]);
+
+            $gallery = new GalleryAll();
+
+            $gallery->image_title = $request->image_title;
+            $gallery->video_title = $request->video_title;
+            if ($request->hasFile('image')) {
+                $image = $request->file('image');
+                $extension = Str::slug($request->image_title) . '-' . Str::random(1) . '.' . $image->getClientOriginalExtension();
+                Image::make($image)->save(public_path('image/' . $extension), 90);
+                $gallery->image = $extension;
+            }
 
 
+            if ($request->hasFile('video')) {
+                $video = $request->file('video');
+                $extension = Str::slug($request->video_title) . '-' . Str::random(1) . '.' . $video->getClientOriginalExtension();
+                $request->video->move('videos', $extension);
+                /* Image::make($video)->save(public_path('video/' . $extension), 90);*/
+                $gallery->video = $extension;
+            }
 
-            // 'regular_price[]' => ['required'],
-        ], [
-            'image.*.mimes' => '  Image must be png,jpg,jpeg formate',
-
-
-        ]);
-
-        $gallery = new GalleryAll;
-
-        $gallery->image_title = $request->image_title;
-        $gallery->video_title = $request->video_title;
-        if ($request->hasFile('image')) {
-            $image = $request->file('image');
-            $extension = Str::slug($request->image_title) . '-' . Str::random(1) . '.' . $image->getClientOriginalExtension();
-            Image::make($image)->save(public_path('image/' . $extension), 90);
-            $gallery->image = $extension;
+            $gallery->save();
+            return redirect()->route('gallery.index')->with('success', 'Added Successfully');
         }
-
-
-        if ($request->hasFile('video')) {
-            $video = $request->file('video');
-            $extension = Str::slug($request->video_title) . '-' . Str::random(1) . '.' . $video->getClientOriginalExtension();
-            $request->video->move('videos', $extension);
-            /* Image::make($video)->save(public_path('video/' . $extension), 90);*/
-            $gallery->video = $extension;
-        }
-
-        $gallery->save();
-        return redirect()->route('gallery.index')->with('success', 'Added Successfully');
+        abort('404');
     }
 
     /**
@@ -99,8 +103,11 @@ class GalleryAllController extends Controller
      */
     public function edit($id)
     {
-        $data['gallery'] = GalleryAll::find($id);
-        return view('backend.gallery.edit', $data);
+        if (auth()->user()->can('Gallery Edit')) {
+            $data['gallery'] = GalleryAll::find($id);
+            return view('backend.gallery.edit', $data);
+        }
+        abort('404');
     }
 
     /**
@@ -112,47 +119,50 @@ class GalleryAllController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
+        if (auth()->user()->can('Gallery Edit')) {
+            $request->validate([
 
 
-            'image' => ['mimes:png,jpeg,jpg',],
-        ], [
-            'image.*.mimes' => '  Image must be png,jpg,jpeg formate',
+                'image' => ['mimes:png,jpeg,jpg',],
+            ], [
+                'image.*.mimes' => '  Image must be png,jpg,jpeg formate',
 
 
-        ]);
+            ]);
 
 
-        $gallery = GalleryAll::findorfail($id);
-        $gallery->image_title = $request->image_title;
-        $gallery->video_title = $request->video_title;
+            $gallery = GalleryAll::findorfail($id);
+            $gallery->image_title = $request->image_title;
+            $gallery->video_title = $request->video_title;
 
-        if ($request->hasFile('image')) {
-            $old_thumbnail = public_path('image/' . $gallery->image);
-            if (file_exists($old_thumbnail)) {
-                unlink($old_thumbnail);
+            if ($request->hasFile('image')) {
+                $old_thumbnail = public_path('image/' . $gallery->image);
+                if (file_exists($old_thumbnail)) {
+                    unlink($old_thumbnail);
+                }
+                $md_thumbnail = $request->file('image');
+                $extension = Str::slug($request->image_title) . '-' . Str::random(1) . '.' . $md_thumbnail->getClientOriginalExtension();
+                Image::make($md_thumbnail)->save(public_path('image/' . $extension), 90);
+
+                $gallery->image = $extension;
             }
-            $md_thumbnail = $request->file('image');
-            $extension = Str::slug($request->image_title) . '-' . Str::random(1) . '.' . $md_thumbnail->getClientOriginalExtension();
-            Image::make($md_thumbnail)->save(public_path('image/' . $extension), 90);
 
-            $gallery->image = $extension;
-        }
-
-        if ($request->hasFile('video')) {
-            $old_video = public_path('videos/' . $gallery->video);
-            if (file_exists($old_video)) {
-                unlink($old_video);
+            if ($request->hasFile('video')) {
+                $old_video = public_path('videos/' . $gallery->video);
+                if (file_exists($old_video)) {
+                    unlink($old_video);
+                }
+                $service_thumbnail = $request->file('video');
+                $extension = Str::slug($request->video_title) . '-' . Str::random(1) . '.' . $service_thumbnail->getClientOriginalExtension();
+                $request->video->move('videos', $extension);
             }
-            $service_thumbnail = $request->file('video');
-            $extension = Str::slug($request->video_title) . '-' . Str::random(1) . '.' . $service_thumbnail->getClientOriginalExtension();
-            $request->video->move('videos', $extension);
+            $gallery->video = $extension;
+
+            $gallery->save();
+
+            return redirect()->route('gallery.index')->with('warning', 'Edited Successfully');
         }
-        $gallery->video = $extension;
-
-        $gallery->save();
-
-        return redirect()->route('gallery.index')->with('warning', 'Edited Successfully');
+        abort('404');
     }
 
     /**
@@ -163,17 +173,20 @@ class GalleryAllController extends Controller
      */
     public function destroy($id)
     {
-        $gallery = GalleryAll::findorfail($id);
-        $old_img = public_path('image/' . $gallery->image);
+        if (auth()->user()->can('Gallery Delete')) {
+            $gallery = GalleryAll::findorfail($id);
+            $old_img = public_path('image/' . $gallery->image);
 
-        @unlink($old_img);
+            @unlink($old_img);
 
 
-        $old_video = public_path('videos/' . $gallery->video);
+            $old_video = public_path('videos/' . $gallery->video);
 
-        @unlink($old_video);
+            @unlink($old_video);
 
-        $gallery->delete();
-        return back()->with('delete', 'Message Deleted Successfully');
+            $gallery->delete();
+            return back()->with('delete', 'Message Deleted Successfully');
+        }
+        abort('404');
     }
 }
