+++ app/Http/Controllers/PagesController.php
@@ -16,10 +16,12 @@ class PagesController extends Controller
      */
     public function index()
     {
-
-        return view('backend.Pages.index', [
-            'pages' => Pages::latest('id')->get(),
-        ]);
+        if (auth()->user()->can('Pages View')) {
+            return view('backend.Pages.index', [
+                'pages' => Pages::latest('id')->get(),
+            ]);
+        }
+        abort('404');
     }
 
     /**
@@ -29,7 +31,10 @@ class PagesController extends Controller
      */
     public function create()
     {
-        return view('backend.Pages.create');
+        if (auth()->user()->can('Pages Create')) {
+            return view('backend.Pages.create');
+        }
+        abort('404');
     }
 
     /**
@@ -40,27 +45,28 @@ class PagesController extends Controller
      */
     public function store(Request $request)
     {
-        // return $request;
-        $request->validate([
-            'title' => ['required'],
-            'heading' => ['required'],
-            'description' => ['required'],
-        ]);
-        $page = new Pages;
-        $page->title = $request->title;
-        $page->slug = Str::slug($request->title);
-        $page->heading = $request->heading;
-        $page->description = $request->description;
+        if (auth()->user()->can('Pages Create')) {
+            $request->validate([
+                'title' => ['required'],
+                'heading' => ['required'],
+                'description' => ['required'],
+            ]);
+            $page = new Pages();
+            $page->title = $request->title;
+            $page->slug = Str::slug($request->title);
+            $page->heading = $request->heading;
+            $page->description = $request->description;
 
-        if ($request->hasFile('thumbnail_img')) {
-
-            $product_thumbnail = $request->file('thumbnail_img');
-            $extension = Str::slug($request->name) . '-' . Str::random(1) . '.' . $product_thumbnail->getClientOriginalExtension();
-            Image::make($product_thumbnail)->save(public_path('thumbnail_img/' . $extension), 90);
-            $page->thumbnail_img = $extension;
+            if ($request->hasFile('thumbnail_img')) {
+                $product_thumbnail = $request->file('thumbnail_img');
+                $extension = Str::slug($request->name) . '-' . Str::random(1) . '.' . $product_thumbnail->getClientOriginalExtension();
+                Image::make($product_thumbnail)->save(public_path('thumbnail_img/' . $extension), 90);
+                $page->thumbnail_img = $extension;
+            }
+            $page->save();
+            return redirect()->route('pages.index')->with('success', 'Page Added Successfully');
         }
-        $page->save();
-        return redirect()->route('pages.index')->with('success', 'Page Added Successfully');
+        abort('404');
     }
 
     /**
@@ -92,14 +98,16 @@ class PagesController extends Controller
      */
     public function edit($id)
     {
-        $pages = Pages::find($id);
-        return view(
-            'backend.Pages.edit',
-
-            [
-                'pages' => $pages,
-            ]
-        );
+        if (auth()->user()->can('Pages Edit')) {
+            $pages = Pages::find($id);
+            return view(
+                'backend.Pages.edit',
+                [
+                    'pages' => $pages,
+                ]
+            );
+        }
+        abort('404');
     }
 
     /**
@@ -111,30 +119,33 @@ class PagesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'title' => ['required'],
-            'heading' => ['required'],
-            'description' => ['required'],
-        ]);
-        $page =  Pages::find($id);
-        $page->title = $request->title;
-        $page->slug = Str::slug($request->title);
-        $page->heading = $request->heading;
-        $page->description = $request->description;
+        if (auth()->user()->can('Pages Edit')) {
+            $request->validate([
+                'title' => ['required'],
+                'heading' => ['required'],
+                'description' => ['required'],
+            ]);
+            $page =  Pages::find($id);
+            $page->title = $request->title;
+            $page->slug = Str::slug($request->title);
+            $page->heading = $request->heading;
+            $page->description = $request->description;
 
-        if ($request->hasFile('thumbnail_img')) {
-            $old_thumbnail = public_path('thumbnail_img/' . $page->thumbnail_img);
-            if (file_exists($old_thumbnail)) {
-                @unlink($old_thumbnail);
+            if ($request->hasFile('thumbnail_img')) {
+                $old_thumbnail = public_path('thumbnail_img/' . $page->thumbnail_img);
+                if (file_exists($old_thumbnail)) {
+                    @unlink($old_thumbnail);
+                }
+                $md_thumbnail = $request->file('thumbnail_img');
+                $extension = Str::slug($request->name) . '-' . Str::random(1) . '.' . $md_thumbnail->getClientOriginalExtension();
+                Image::make($md_thumbnail)->save(public_path('thumbnail_img/' . $extension), 90);
+
+                $page->thumbnail_img = $extension;
             }
-            $md_thumbnail = $request->file('thumbnail_img');
-            $extension = Str::slug($request->name) . '-' . Str::random(1) . '.' . $md_thumbnail->getClientOriginalExtension();
-            Image::make($md_thumbnail)->save(public_path('thumbnail_img/' . $extension), 90);
-
-            $page->thumbnail_img = $extension;
+            $page->save();
+            return redirect()->route('pages.index')->with('success', 'Page Eidted Successfully');
         }
-        $page->save();
-        return redirect()->route('pages.index')->with('success', 'Page Eidted Successfully');
+        abort('404');
     }
 
     /**
@@ -145,7 +156,10 @@ class PagesController extends Controller
      */
     public function destroy(Pages $pages, $id)
     {
-        Pages::find($id)->delete();
-        return redirect()->route('pages.index')->with('delete', 'Page Deleted');
+        if (auth()->user()->can('Pages Delete')) {
+            Pages::find($id)->delete();
+            return redirect()->route('pages.index')->with('delete', 'Page Deleted');
+        }
+        abort('404');
     }
 }
